Move package.json & README to library src
