--- js/npm/jsonrpc/README.md
@@ -1,5 +1,0 @@
-# @parity/jsonrpc
-
-JSON and JS interface defintions for RPC calls.
-
-[https://github.com/paritytech/parity/tree/master/js/src/jsonrpc](https://github.com/paritytech/parity/tree/master/js/src/jsonrpc)
