--- js/npm/jsonrpc/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@parity/jsonrpc",
-  "description": "JSON and JS interface defintions for RPC",
-  "version": "0.0.0",
-  "main": "library.js",
-  "author": "Parity Team <admin@parity.io>",
-  "maintainers": [
-    "Jaco Greeff"
-  ],
-  "contributors": [],
-  "license": "GPL-3.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/paritytech/parity.git"
-  },
-  "keywords": [
-    "Ethereum",
-    "ABI",
-    "API",
-    "RPC",
-    "Parity"
-  ],
-  "scripts": {
-  },
-  "devDependencies": {
-  },
-  "dependencies": {
-  }
-}
