+++ js/package.json
@@ -163,7 +163,10 @@
     "@parity/abi": "file:packages/abi",
     "@parity/api": "file:packages/api",
     "@parity/dapps": "file:src/dapps",
+    "@parity/etherscan": "file:packages/etherscan",
     "@parity/jsonrpc": "file:packages/jsonrpc",
+    "@parity/ledger": "file:packages/ledger",
+    "@parity/shapeshift": "file:packages/shapeshift",
     "@parity/shared": "file:packages/shared",
     "@parity/shell": "file:src/shell",
     "@parity/ui": "file:packages/ui",
