+++ js/packages/jsonrpc/library.js
@@ -14,20 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import es6Promise from 'es6-promise';
-es6Promise.polyfill();
+import JsonRpc from './';
 
-const isNode = typeof global !== 'undefined' && typeof global !== 'undefined';
-const isBrowser = typeof self !== 'undefined' && typeof self.window !== 'undefined';
-
-if (isBrowser) {
-  require('whatwg-fetch');
-}
-
-if (isNode) {
-  global.fetch = require('node-fetch');
-}
-
-import Etherscan from './3rdparty/etherscan';
-
-export default Etherscan;
+export default JsonRpc;
