+++ js/packages/shapeshift/library.js
@@ -14,6 +14,20 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import JsonRpc from './jsonrpc';
+import es6Promise from 'es6-promise';
+es6Promise.polyfill();
 
-export default JsonRpc;
+const isNode = typeof global !== 'undefined' && typeof global !== 'undefined';
+const isBrowser = typeof self !== 'undefined' && typeof self.window !== 'undefined';
+
+if (isBrowser) {
+  require('whatwg-fetch');
+}
+
+if (isNode) {
+  global.fetch = require('node-fetch');
+}
+
+import ShapeShift from './';
+
+export default ShapeShift;
