+++ js/packages/ui/TxHash/txHash.js
@@ -21,8 +21,8 @@ import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 
 import { DEFAULT_GAS } from '@parity/shared/util/constants';
+import { txLink } from '@parity/etherscan/links';
 
-import { txLink } from '~/3rdparty/etherscan/links';
 import Warning from '~/ui/Warning';
 import Progress from '~/ui/Progress';
 import ShortenedHash from '~/ui/ShortenedHash';
