+++ js/src/views/Account/Transactions/store.js
@@ -16,7 +16,7 @@
 
 import { action, observable, transaction } from 'mobx';
 
-import etherscan from '~/3rdparty/etherscan';
+import etherscan from '@parity/etherscan';
 
 export default class Store {
   @observable address = null;
