+++ js/src/views/Account/Transactions/store.spec.js
@@ -17,7 +17,7 @@
 import BigNumber from 'bignumber.js';
 import sinon from 'sinon';
 
-import { mockget as mockEtherscan } from '~/3rdparty/etherscan/helpers.spec.js';
+import { mockget as mockEtherscan } from '@parity/etherscan/helpers.spec.js';
 import { ADDRESS, createApi } from './transactions.test.js';
 
 import Store from './store';
